Merged PR 16: Add PackedMap based on Viktor Leis' B Tree node
This PR checks in Viktor Leis' B Tree node code. I've placed the node into a `PackedMap` class that is templated with a size so that we have the option to use multiple differently sized maps simultaneously in the codebase.

I did not use our `Slice` abstraction here so that this code remains "independent" from the rest of our codebase.

Most of the code is unchanged from Viktor's implementation. I did need to make two changes:

**1. Initial Fences:** We need to specify the map's key range when constructing the map to take advantage of prefixed encoding. I believe the B Tree handles the fences when splitting/merging nodes, which we do not have in our use case. The relevant code changes are in the new non-default `PackedMap(...)` constructor.

**2. Duplicate Inserts:** Duplicate inserts with the same key should overwrite the associated value. However, Viktor's node handled duplicate inserts with the assumption that the record value's length is unchanged. I don't think we can make this assumption in our case so I've modified `Insert(...)` to handle this case. I needed to add a separate heap compaction method to handle this case as well: `RemoveFromHeapAndCompact(...)`.

I've added some tests so we can be more confident in the implementation's correctness (see `BoundedOverwrite` and `OverwriteCompact`).