Merged PR 17: Add Page abstraction to manage the on-disk format
This adds a `Page` class that wraps the `PackedMap` class. The idea would be that we use this class when modifying the page in memory.

The encoding/loading API is not finalized (in particular `EncodeTo()`). The LevelDB codebase uses `std::string` as a variable length byte buffer. We may want to do something different later on to reduce the number of times we need to copy memory.