Merged PR 18: Add initial MemTable implementation based on std::map
This PR adds a `MemTable` that will be the initial "parking place" for writes that come in to the database. As discussed, we initially use a `std::map`. It should be relatively straightforward to replace this with something else later on (e.g., a learned index).

We need to explicitly store deletes in the `MemTable` so that we can delete the keys from the on-disk page when flushing. We also need this information for the read path to avoid loading the page from disk.

To store the keys and values, I also checked in LevelDB's `Arena` implementation. The way they use their `MemTable` is that they delete and re-construct it after flushing. We will need to take a similar approach (or will have to reset the `Arena` after each flush). So `arena.h` and `arena.cc` have been copied over from LevelDB.