Merged PR 23: Update the Page class so that it uses a separate memory buffer
This PR updates the `Page` class so that it no longer "owns" the memory where its data is stored, but instead uses a pointer to custom-managed memory. The idea is that this memory would be managed by the buffer manager.

This also sets our page size to ~64 KiB (see comments in the source code).