Merged PR 27: Make record size configurable
These changes allow us to configure the size of a record in the `synth_write` benchmark. The main change is to `U64Dataset` in `data.h`/`data.cc`, which now holds a byte buffer that stores all the values contiguously. The keys are still generated in the same way as before.