Merged PR 28: Allow segments and pages per segment to not be powers of two
When setting up the initial benchmarks, I found that requiring the number of pages per segment to be a power of 2 could result in needing a lot of empty pages. For example, if we want to write 512 MiB while filling the DB pages to just under 50%, the total size of the DB could end up at 2 GiB.

This PR changes the page and segment addressing scheme to instead allow the segments and pages per segment to not necessarily be powers of two.

# Microbenchmark results (dataset size)

## Large dataset, large records
512 MiB User Data, 512 B Records

**Before:** 2 GiB Total Database Size
**After:** 1.2 GiB

## Small dataset, small records
64 MiB User Data,  16 B Records

**Before:** 256 MiB Total Database Size
**After:** 143 MiB