Merged PR 30: Implement initial LLSM synthetic write benchmark
This PR implements the initial `synth_write` bindings for LLSM. This is a first pass implementation. In particular, LLSM currently does not automatically flush the memtable when it is "full" (it has to be triggered by the user). This also means that the flush does not happen in the background, so it currently does not overlap any memtable writes. Nevertheless, this is enough to start getting some performance numbers so we know where we stand relative to RocksDB.