Merged PR 36: Separate out Model Class
Created `Model`, a separate wrapper class for the models, alongside `DirectModel`, the pseudo-model we have been using so far.

Many parts of the current design are inter-connected, so here is how they interact at present (you can also follow this in `db_impl.cc`:
1. The database starts out with some user-defined database options in an `Options` struct.
2. The `DirectModel` constructor is called and given the `Options` struct. This constructor makes all the necessary calculations (how many records per page, pages per segment etc.) and sets a `BufMgrOptions` struct accordingly - it also includes a reference to itself so that the `FileManager` can later get access to the model.
3. The `BufferManager` constructor is called and given the `BufMgrOptions` struct. We couldn't do this before step 2 because the `FileManager` inside the `BufferManager` needs access to the model (to calculate the `FileAddress` from the page_id).
4. The `DirectModel::Preallocate()` method is called to preallocate the right number of pages. We couldn't do this before step 3 because the model needs a functional `BufferManager` to fix/unfix pages during preallocation.

I opened the pull request at this point so that we can go over this sequence and the `Model` API before more elaborate models are considered. For example, should `PageIdToAddress()` be part of the model or the `FileManager`? The first option (currently implemented) gives us potentially more flexibility when designing what to do with overflow pages, whereas the second one (previously implemented) eliminates the dependence of the `FileManager` (and, by extension, of the `BufferManager`) on the model.