Merged PR 37: Switch to using RocksDB's skip list
This PR brings in RocksDB's skip list and makes the necessary changes in `memtable.h/.cc` to get it integrated.

The main changes to review are in `memtable.h` and `memtable.cc`. One difference between LevelDB's skip list and RocksDB's skip list is that RocksDB's skip list (`InlineSkipList`) works with raw/encoded records. So the main changes in our `memtable.h/.cc` files are to deal with that change.

The files copied over from RocksDB are:
- `inlineskiplist.h`
- `random.h/.cc`
- `gcc_macros.h` (selectively adapted from RocksDB's headers)

I did also need to make a few minor changes to the files listed above to get them integrated with our codebase. To see the specific changes, you can look at the "Untangle from RocksDB code" commit.