Merged PR 43: Add a more robust and general purpose thread pool
This PR adds a more robust and general purpose thread pool that we can use in LLSM. I looked at several thread pool implementations (listed in the comments above `class ThreadPool`) and none of them were that well suited for our codebase. So instead this custom thread pool implementation combines the "best parts" of the existing implementations.

There are two `Submit()` methods because I was concerned that the one that returns a future would be more expensive than without. The microbenchmarks show that the two methods aren't that different in overhead (at 16 threads, it's about 2 to 3 us per submit). Nevertheless, it might be useful to keep both methods for when we don't want the future returned.