Merged PR 45: Make DB thread-safe and implement background memtable flushing
This PR adds locking to `DBImpl` to make its interface thread safe. At the same time this also adds automated memtable flushing (in the background) with double buffering.

The primary changes are in `db_impl.h` and `db_impl.cc`. I added comments to explain the changes and also added comments above the instance variables in `DBImpl` to indicate whether or not the mutex needs to be held to access the variable.

The main tricky part of these changes is handling full memtables. In the situation where a flush is in progress *and* the active memtable is full, we need to block additional writers until the flush finishes (so there is space to do the write). However we still want readers to proceed. So the `waiting_writers_` queue is used to queue up incoming writer threads while the flush is pending.