Merged PR 46: Implement Get() and add simple end-to-end test
This PR adds a first pass thread-safe `Get()` implementation. We check the memtables first and then use the model to read a disk page if needed. I also added a very simple end-to-end functionality test.

We'll probably need to add something more sophisticated later to check for end-to-end correctness, but I think that can wait until we get around to implementing loading an existing LLSM database from disk.