Merged PR 49: Add search-free `Append()` functionality to `PackedMap`
Added a new `Append()` function to `PackedMap`, that bypasses the binary search among the slots and tries to insert directly at the end. The function calls `Insert()` instead if the provided `key` is not greater than the largest key currently in the `PackedMap`.

This check can be bypassed by setting the `perform_checks` parameter to `false`, but the user needs to guarantee that the appended keys will be sorted. We could add a repair process that re-sorts the `Slots` if the user violates that guarantee, but it seems to me that the current set-up is good enough for our current bulk-loading use case (an unsure user can simply choose to always have `perform_checks = true`). Misusing `perform_checks = false` would cost `O(nlogn)` anyway, because of the sort we would later need, which is also the cost of just using `Insert()` instead (`O(logn)` for each of `n` keys).

I also changed the address calculation in the `FileManager`, which now uses binary instead of linear search to find the right segment. This is certainly more efficient algorithmically for large numbers of segments, but I'm not sure it will make a real difference with only a few of them.