Merged PR 51: Allow concurrent memtable reads, exit on full page write attempt
This PR modifies the DB's read path to allow readers to search the active memtable without holding the database lock. The RocksDB skip list allows reads to proceed concurrently with writes. We still need to hold the mutex during writes because we use the skip list's `Insert()` method, which requires external mutual exclusion.

We may be able to get further performance benefits if we use the `std::atomic*` functions that help with copying `shared_ptr`s. I think we should leave that to later because it would require changes in `WriteImpl()` too.

I also modified the flush worker thread to call `exit()` if it needs to write to a full page. This is just to prevent silent failures in our benchmark until we support overflow pages.