Merged PR 52: Move common LLSM/RocksDB benchmark options into config files
This moves the common configuration flags used in `synth_write` that are also used in the YCSB benchmark driver (added in a different PR) into a configuration header. This allows us to share the same flags across the benchmarks and also centralizes the code that sets the databases' options structs to ensure they are set consistently.

This also modifies the `synth_write` benchmark so that RocksDB always does a compaction after the bulk load (the compaction run time is counted towards the benchmark's overall run time).

Related work items: #4