Merged PR 53: Add YCSB benchmark driver
This PR adds in a YCSB benchmark driver that runs the workload on LLSM and RocksDB.

The way the benchmark driver works is that it uses a "workload file", which is just a trace of the operations the official YCSB benchmark driver makes. The trace* contains the operation (insert, update, read, scan) and the key (a 64-bit integer). The values are generated at runtime and the value size is runtime-configurable.

The benchmark code is based on our earlier YCSB benchmarking code. The code is reusable, so I packaged it as a header-only library under the `ycsbr` namespace and pushed it to a separate repository called `ycsb-runner`. The library is included as a dependency under `third_party`.

*To extract a trace, there's a tool in the `ycsb-runner` repository that consumes the output of the official YCSB benchmark driver and outputs a serialized trace file that can be used by the `ycsbr` code.

Related work items: #4