Merged PR 54: Ensure buffers are aligned to the file system's block size
The root cause of our direct I/O issue is that our memory buffer is not aligned to the file system's block size.

It turns out that in ext4, writing threads hold an exclusive lock on a file inode if (i) the write requires metadata changes (e.g., the file grows in size, new blocks need to be allocated, etc.) or (ii) the write is not aligned to the file system's block size.

We are not affected by (i) because we preallocate the segments. However we *are* affected by (ii). It turns out that the write memory buffer *also* has to be aligned to a multiple of the file system's block size. Apparently just having the file offsets being aligned to the file system's block size was not enough.

The relevant parts of the kernel code are here (we're running v5.9.1 on `berners-lee` and `mcgraw`):

- https://elixir.bootlin.com/linux/v5.9.1/source/fs/ext4/file.c#L468
- https://elixir.bootlin.com/linux/v5.9.1/source/fs/ext4/file.c#L404
- https://elixir.bootlin.com/linux/v5.9.1/source/lib/iov_iter.c#L1225

With our memory buffer aligned to the file system's block size, our writes can proceed by just holding a shared lock on the file inode, which allows the writes to proceed in parallel.

---

This PR changes the buffer manager's memory alignment code so that we automatically set it to the underlying file system's block size by using `statvfs()`. If for some reason the call fails, we set the alignment to 4096.

These changes mainly affect the sorted load `synth_write` benchmark. When writing a 4 GiB dataset of 512 B records using 16 background threads and 50% filled pages, we now achieve a user data write throughput of 338 MiB/s (versus 172 MiB/s without these changes). Now we may also want to revisit having a segment per thread.