Merged PR 57: Refactor: Move EntryType out of MemTable
This PR moves the `EntryType` enum out of the `MemTable` and into a separate format header. This is because we need to also disambiguate between writes and deletes in the write-ahead log, so I think it makes sense to share the same enum between the memtable and log.

We're using the `format.h` name to be consistent with LevelDB and RocksDB, which also uses the same header to list numeric values that will be used in the on-disk files.