Merged PR 58: Add write-ahead log reader and writer
This PR adds reader and writer classes that help with reading from and writing to the write-ahead log respectively. This PR also includes encoding utilities and adds a dependency on CRC32C for checksums that are used in the log. These classes will be built with LLSM, but currently are not used (will come in a later PR).

The idea is that these classes are not supposed to be used independently, but instead through a manager class that offers a higher level interface. This is because we need to manage multiple versions of the log files, both during recovery and for keeping track of when a file can be deleted. I'm still making adjustments to the manager code based on how we will integrate it with `DBImpl`.

**Code coming from LevelDB (without modification):**
- `util/crc32c.h` (Calls into the CRC32C library to compute CRC32C checksums)
- `util/coding.h` (Encoding utilities, including variable length ints, which could be useful later on as well.)
- `tests/coding_test.cc`

**Code coming from LevelDB (with modifications):**
- `wal/reader.h|.cc`
  - See the "Modify reader.h/.cc to work with LLSM" commit for the changes
- `wal/writer.h|.cc`
  - See the "Modify writer.h/.cc to work with LLSM" commit for the changes
- `format.h` (Added the documentation and switched to `enum class`es)
- `tests/wal_rw_test.cc` (Adjustments to the tests to work with our changes)

**Other changes:**
- Added a `Status` helper to report `errno` error codes from system calls
- Added a dependency on Google's CRC32C library (fast implementations of CRC32C)
- Added options to `WriteOptions` to give users control over the log's sync semantics (same as available on RocksDB)