Merged PR 59: Avoid overly blocking reads during a write
Right now because reading threads need to acquire `mutex_` to copy the memtable shared pointers, they need to wait while a write takes place. This is not that ideal because the skip list supports reading while writes are taking place. This will become a bigger problem when we add in the WAL because it may slow down the write path (especially if the log is fsync'd).

This PR adds a second mutex used specifically to protect the memtable shared pointers. Because of how our write path is set up, a writing thread only needs to acquire this mutex when updating the memtable shared pointers (when a flush is scheduled). This allows reads to occur concurrently with writes.

---

I sanity checked the performance with `synth_write` (to make sure pure writes aren't negatively affected) and YCSB A with 4 application threads. The `synth_write` numbers look the same. For YCSB A, we get a slight improvement for the 16 B record experiment (was: 48.9 kops/s, now: 56.0 kops/s, approx. 1.3x better than RocksDB in the same configuration). There doesn't seem to be a difference for large records. These numbers are without deferred I/O running.