Merged PR 60: Add write-ahead log microbenchmark
This PR adds a microbenchmark that measures:

- Making single writes to the log with and without `fdatasync`
- Making batched writes to the log with and without `fdatasync`
- Writing to the memtable (single writes and batched writes)
- Writing to the memtable and the log (single writes and batched writes), with and without `fdatasync`
- Calling `fdatasync()` on a file descriptor after having made a write of some fixed size

The reason for including batched writes is that LevelDB/RocksDB both support batched writes as well. The reason for putting this microbenchmark in its own executable is to allow users to pass in a desired output path for the log. This helps make it easier to run the benchmark on different storage devices.

In the example run I only included a few of the results, since there are lots of configurations that the microbenchmark runs.