Merged PR 61: Add the WAL manager and integrate it with LLSM
The main part of this PR is the `wal::Manager` class, which provides a high-level interface over the `wal::Reader` and `wal::Writer` classes. The idea is for it to help with transparently managing the log versions. The rest of the changes are for integrating the manager with `DBImpl` and the benchmark drivers. So the main files to look over are `manager.h|.cc` and `db_impl.cc`.

The high level idea behind managing the logs is as follows:
- We store the logs under a separate `wal` subdirectory.
- The name of each log file is its version number.
- Version numbers start from 0 and increase monotonically.
- Each log version is associated with a memtable "instance/generation". Once a memtable is sent to be flushed, we move on to the next log version.
- After a flush completes, we delete the oldest log version while maintaining enough old versions based on the maximum number of times a record can be deferred.
- On a clean shutdown, we discard all the logs after flushing the memtable.

A few other implementation details:
- I set up the manager to lazily create the underlying `wal::Writer` (i.e., it waits until the first log write to actually create the `wal::Writer`, which then creates the log file). This is to avoid creating files when all the writes bypass the WAL.
- I tried to mostly use the standard library's `<filesystem>` utilities to do the file creation/deletion so that the implementation is a bit easier to read.
- We always `fsync()` the `wal` directory when we delete files.
- On log creation, we `fsync()` the directory the first time we get a write where `options.sync == true`.

Note that this PR does not implement "re-opening" an existing DB. I plan to work on that next as it will help with running our benchmarks.