Merged PR 62: Various tech debt fixes in preparation for reopening
This PR touches a lot of files, but it contains just a few high-level refactoring changes to the code. These changes are meant to address some of the outstanding tech debt in the codebase to prepare for the reopening changes.

The high-level changes are:

1. Replaced our usage of `std::string` with `std::filesystem::path` when dealing with paths.
2. Fixed the comments above `Get()` so that they refer to the new mutex from last week.
3. Moved the options that deal with describing the key distribution into a separate struct called `KeyDistHints` (other name suggestions are welcomed!)
4. Refactored `DBImpl::Initialize()` to split out the "new DB" and "existing DB" initialization paths.
5. Fixed a minor bug - prevent an assertion failure when `~DBImpl()` is called on an uninitialized `DBImpl`.

The reasons for these changes are:

For (1): I think it's semantically clearer to use `std::filesystem::path` to refer to a path; it's a C++17 feature, but we're already compiling against C++17 anyways. Also, `std::string` and C-strings appear to both be implicitly convertible to `path`. So this change will not affect any client code.

For (2): Just making sure the comments match the code. I forgot to change them in the last PR.

For (3): The options in `KeyDistHints` will only initially be used when creating a new DB. I think it will help to keep them defined in a struct that is separate from the rest of the top-level `llsm::Options` struct.

For (4): The plan is to have `InitializeExistingDB()` read `MANIFEST` to restore the DB state. Then `InitializeNewDB()` will only be used when creating a new DB, and it will contain the existing set up code.

For (5): If there's an initialization failure, we delete the `DBImpl`. But this calls `wal_.DiscardAllForCleanShutdown()` before the WAL manager is set to write mode, which causes an assertion failure. I think a reasonable way to fix this is to just allow calls of this method when the manager is in "created" mode (and make it a no-op).

Related work items: #2