Merged PR 63: Add de/serialization to Model-related classes
This PR adds `EncodeTo()` and `LoadFrom()` to the `Model` class to support serialization/deserialization. The serialization and deserialization is done by writing to a `std::string` which acts as a byte buffer. The changes in !64 use these methods to actually write the serialized data to a file.

Related work items: #2