Merged PR 64: Add Manifest related files
This PR adds a `Manifest` class (and an associated builder class) to encapsulate the logic needed for serializing and deserializing the manifest. The file will be a binary file; I've left some comments in `manifest.cc` describing its current layout.

The manifest currently saves (and restores):
- The number of pages in the database
- The number of segments used
- The model

The database code will recompute the number of pages per segment by assuming that the pages are equally divided among the segments. If this assumption changes in the future we will just need to store more metadata in the manifest.

Related work items: #2