Merged PR 65: Reintroduce BufMgrOptions
This PR adds `BufMgrOptions` back in to allow for configuring the `BufferManager` and `FileManager` on construction. This PR also modifies `KeyDistHints` so that `records_per_page` is now a method that computes the number of records per page (based on the other hint options). This helps us avoid duplicating the code that computes this value.

Related work items: #2