Merged PR 66: Implement database reopening using the manifest
This PR mainly updates `db_impl.cc` so that we actually use the code added in the previous PRs to serialize the manifest. This PR also includes logic for deserializing the manifest and preparing the database in `InitializeExistingDB()`.

I also modified the WAL manager's replay behavior to allow for an "early abort" to the replay if a callback ever returns a non-OK status. This allows us to abort the replay early if any of the writes fail.

Related work items: #2