Merged PR 68: Include Valgrind in our Azure Pipeline
Based on our discussion on Tuesday, I looked into including Valgrind in our Azure Pipeline. This PR contains changes to `azure_pipelines.yml` which adds a "Valgrind Memcheck" step at the end of the build and test job. I set up some additional configuration options to let us skip the tests and/or Valgrind, which might be useful when launching the pipeline manually for debugging purposes.

Based on follow up discussions during code review, this PR also adds in a second pipeline that we will run on pull requests only. The pull request pipeline will only run Valgrind Memcheck on `DBTest.*` tests, which speeds up the time to pipeline completion. When we commit to `master`, we will run the checks again but this time against the whole test suite.

Valgrind also caught a few existing memory errors and so I included fixes for them in this PR as well (they were minor errors and I would not expect them to affect performance).