Merged PR 74: Keep track of file sizes when reopening
It turns out that the missing reads issue I mentioned during our Tuesday meeting was related to DB reopening, but not to the serialization.

The `File` class keeps track of the file size in order to zero out ranges in the on-disk file when writing to new offsets in the file. The problem was that we always set this file size tracking variable to 0 when creating the `File`, and then use `ftruncate()` to resize the on-disk file as needed. If the file already exists, and an "early" page is accessed, this would cause the `File` class to shrink the file on disk and essentially throw away pages. This was why reading from a reopened database led to lost writes.

To fix the bug I modified `File` to load the file size using `fstat()` when it is constructed. I also switched `ftruncate()` to `fallocate()` just to be extra safe since `fallocate()` with mode `0` will never shrink a file. One downside however is that `fallocate()` is a Linux-specific system call (I think we could also use `posix_fallocate()` for something more portable, if that's needed).

I also added a new test to `db_test.cc` that reproduces this bug (I verified that it fails on `master`).