Merged PR 77: Flush statistics
This PR adds variables to `db_impl.h/.cc` to track the number of flushes, buffer manager hits, IOs and deferrals, as well as the number of records associated with each of these events. This feature is activated by setting `options.flush_stats=true` when creating the database.

This PR also adds/renames two convenience scripts:
*  `scripts/run_ycsb_w_deferral.sh` for running `bench/ycsb` for different settings of the deferral parameters.
* `scripts/run_deferred_io.sh` for running `bench/deferred_io` for different settings of the deferral parameters.