Merged PR 78: Add cache size to the automated YCSB benchmarks
When putting together the experiment definitions in !75, I forgot to set the cache size in the YCSB benchmarks. This means the default size of 64 MiB was being used.

The values are from https://dev.azure.com/msr-dsail/LearnedLSM/_wiki/wikis/LearnedLSM.wiki/3/Benchmarks. The cache size is around 33% of the total dataset size. Note that this is larger than the size we used in the initial YCSB experiments (originally ~17% of the total dataset size).