Merged PR 80: Add a RocksDB comparison to the automated experiments
This PR modifies our automated experiment infrastructure to also include a comparison against RocksDB. The RocksDB experiments are only run if their results do not exist (i.e., after they run once, we cache the results and reuse them for later comparisons).

I also modified the email to highlight the change/speedup columns. The number will be highlighted red if it is less than 0.95 and green if it is greater than 1.05. Put another way, any change +/- 5% is not highlighted.