Merged PR 81: Make pages iterable
To prepare for adding support for range scans, this PR adds the ability to iterate over the records stored in a page. This PR also exposes the slot abstraction in `PackedMap`s (previously you could only get records by key), which is needed to support in-order iteration. I think that using an iterator interface is reasonable for `Page`s since they will be in memory when we iterate over them.

Markos had also pointed out that we're currently using `Page::kSize` to compute the number of records per page even though some of the page is used for metadata (the page header). Since these changes already touch the `Page` class and `PackedMap`, I also went ahead and modified the classes to expose the number of "usable bytes" in the page (usable for storing data). This is essentially the page size minus the size of the page header. The `Page` class now also exposes the size of the "slots" used for each record.

Related work items: #8