Merged PR 83: Introduce a Page Table
This PR introduces a page table into the `FileManager`.

It switches to a full-on `FileAddress` class to halve the memory requirements by concatenating the `file_id` and `offset` into a single variable.

Support for updating the page table has not been added yet, as it will depend on the overflow strategy.

Related work items: #12