Merged PR 84: First pass at range scans
This PR contains a first pass at implementing range scans, and is the code that I used to get our initial results from last Thursday's meeting. This is a simple implementation with no prefetching - we probably will want to add prefetching in later on to help with longer scans. This implementation also assumes that:

- There are a fixed number of pages
- There are no page overflows
- We can simply increment the page ID to get the next page of records

These assumptions of course will not be valid after we add in the reorganization. I was thinking we could commit these changes for now and then update the range scan code as we make the reorganization changes.

The API is batch-based and asks for a starting key and the maximum number of records to fetch. The reason for choosing this interface right now is that it is the format that YCSB uses to make requests (start key and the number of records to scan). Eventually we may need to add in a `[start key, end key]` based API method, and even an iterator.

Since the `db_impl.cc` file was getting long, I created a new `db_impl_range.cc` file to hold the range scan related `DBImpl` code.

Related work items: #8