Merged PR 85: [Overflow 1/4] Read path can handle overflow
This PR modifies the read path in `db_impl.cc` so that it checks the next linked overflow page, if any, in the cases where the lookup was not successful in the current page.

The convention used is that overflow pages have `page_id`'s with 1 as the most significant bit, so the `overflow` field in a `PackedMap` is only really valid when this condition is met (e.g. it's not valid upon page construction, since this field is initialized to 0).

This saves an extra flag variable by limiting the potential number of "regular" pages by a factor of 2, but 2^63 should still be plenty.

Related work items: #15