Merged PR 94: Add aligned memory allocation helper
This PR adds a `PageMemoryAllocator` class to help with allocating aligned memory that is meant to be used to store pages read/written from/to persistent storage using direct I/O. I thought it might be helpful to have the allocator encapsulated in a class because we allocate pages in memory in several places (and may need to do so during reorganization too). This class does not need to be used in our tests because we do not use direct I/O there.

The allocator uses static storage to hold the alignment that should be used. This simplifies the code because we won't need to pass around a class instance to different parts of the code that need to allocate a page. However the downside is that only one DB instance can be used at any time (for absolute correctness). I think this limitation is reasonable for our use case since we also already use static storage in `PhysicalPageId`.

I also disabled Valgrind for now because of the unsupported instructions.