Merged PR 95: Add BTree-based model
This PR adds `BTreeModel` using largely the same implementation as `ALEXModel`. It is based on the TLX `btree_map`, which is the current version of the now-obsolete STX `btree_map`.

An option to toggle between the two models and a corresponding flag have also been added for ease of benchmarking.

Related work items: #36