Merged PR 98: Check in reusable timing code from cuckoo-index
This PR checks in the profiler from https://github.com/google/cuckoo-index/blob/master/common/profiling.h and removes the dependencies on abseil. I also did some renaming since the class is more of a timer than a profiler.

The class is not used anywhere. The intention is that we will add it in locally when we need to debug parts of the code.

Related work items: #47