Merged PR 100: Deferral batch size autotuning
This PR implements one of the autotuning techniques that @<Geoffrey Yu>  and I proposed for out 6.830 class project - namely, the system automatically selects the deferral threshold.

The threshold variable has been changed to represent the size of a batch in bytes, rather than the number of records, to accommodate variable-sized records.

The simple cost model used to pick the threshold is as follows:

 $batch\_size = sf * \sqrt{\frac{memtable\_size * page\_size * BUFMGR\_AVG\_MISS\_TIME}{MEMTABLE\_FILL\_TIME}}$

where:
* $sf$ is an appropriate constant hyperparameter, which is currently set to 1. We might not need this knob after all, but it could provide some extra flexibility.
* $memtable\_size$, in bytes, plays a role because a larger memtable can absorb larger deferred batches.
* $page\_size$, in bytes, plays a role because larger pages increase the write amplification for batches of a given size, so deferral becomes more attractive.
* $BUFMGR\_AVG\_MISS\_TIME$ is the average latency associated with a buffer manager miss. The larger it is, the more we gain by deferring.
 * $MEMTABLE\_FILL\_TIME$ is the time between allocation and flush for the most recent memtable instance, which indicates the rate of user inserts. The higher this rate is, the less we should defer, to avoid occupying memtable space with deferred batches.

Right now on YCSB A this seems to perform within 2% of the optimal manually-set deferral batch size, but its real usefulness will be on evolving workloads, where different batch sizes will be appropriate for each phase.

What this technique doesn't touch is the maximum number of deferrals, since it is currently related to our WAL implementation. This goes back to a discussion with Umar about disconnecting the log truncation from the flush process, which we haven't yet implemented. The default value of `deferred_io_max_deferrals` is changed to 1, however, so that this option does not need to always be set when setting `deferral_autotuning=true`.

Related work items: #23, #38