Merged PR 101: Fixes to support workloads with inserts
This PR adds some extra checks to the reorganization and range scan code to handle the case when `FixOverflowChain()` returns `nullptr`. This was one of the bugs that was causing errors when running YCSB E (95% range scans, 5% inserts).

The other bugs were in ALEX. I made two changes (see the comments) that seem to be enough to allow our current YCSB E (synthetic data) workload to run to completion. Both bugs were a result of signed integer overflow.

An underlying problem is that the ALEX code uses signed 32-bit ints throughout the codebase. The code usually assumes the values are positive, so there can be problems if the ints wraparound. We'll need to keep an eye out for this problem since I assume it may show up in other parts of the code too.