Merged PR 103: Switch to 4 KiB pages
This switches our page size to 4 KiB, based on our discussion from Tuesday. I updated the tests to make them dependent on the `Page::kSize` constant. So it should be simple to switch again if needed (i.e., future page size changes up to 64 KiB should be a 1 line change).