Merged PR 108: Adjustable memtable flush threshold, check in autotuning experiments
For 6.830, we triggered memory autotuning every 1 million requests. To support that change, we also needed to be able to modify the flush threshold of an existing memtable. I removed the 1 million request trigger, but thought it would be helpful to keep the modifiable flush threshold since we will add in triggering again later on anyways. So the changes under `db/` are primarily to support the modifiable flush threshold.

The rest of the changes under `scripts` are our experiments. I did some refactoring for `e2e_perf` and also included the sweep experiment (0%-100% writes in 10% increments).

Related work items: #26