Merged PR 111: YCSB with custom datasets, unfix pages if needed during reorganization
This PR includes the YCSB experiments with custom datasets (currently only A, B, C, F are enabled, and we use the BTree model by default for now). It also includes some changes I made to LLSM to actually use the B-tree and when investigating the issue with inserts.

Related work items: #52, #53