Back off when fixing chains if there are no frames left (#4)
This PR contains the changes I described in our previous meetings: if there are no frames left when fixing a chain, the thread doing the fixing will "back off" and retry. Doing this gives other threads an opportunity to successfully fix their entire chain. The upside of this approach is that it adds minimal overhead to the "common" case, which is when there are enough frames to fix the entire chain. The downside is that a livelock is still theoretically possible if all the threads fixing the chains proceed in perfect unison, which should be unlikely.

I tested these changes on the OSM dataset and we are now able to load the dataset into LLSM. Before, we would run into a livelock situation where all the background threads would end up spinning while waiting for a frame to become available.